backend/requirements.txt
python-socketio == 5.11.4
eventlet == 0.38.0
pydealer